Support decimal values for quantity

+++ GateIo.Net/Objects/Models/GateIoPerpLiquidation.cs
@@ -29,7 +29,7 @@ namespace GateIo.Net.Objects.Models
         /// ["<c>size</c>"] Quantity
         /// </summary>
         [JsonPropertyName("size")]
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         /// <summary>
         /// ["<c>leverage</c>"] Leverage
         /// </summary>
